Get rid of signed/unsigned comparison warning in assert
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3481 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/uqm/oscill.c
@@ -76,7 +76,7 @@ DrawOscilloscope (void)
 	if (oscillDisabled)
 		return;
 
-	assert (scopeSize.width <= sizeof scope_data);
+	assert ((size_t)scopeSize.width <= sizeof scope_data);
 	assert (scopeSize.height < 256);
 
 	if (GraphForegroundStream (scope_data, scopeSize.width, scopeSize.height))
